Move functions higher up namespace
Move gen_distribution() and gen_repo() up the namespace hierarchy. This
lets tests for different plugins avoid importing from each other. In
this particular case, file plugin tests no longer need to import from
the pulp core tests. This is especially useful if plugins are going to
be pulled out of Pulp Smash and pushed into individual plugins.
