+++ pulp_smash/tests/pulp3/file/api_v3/test_auto_distribution.py
@@ -15,8 +15,9 @@ from pulp_smash.tests.pulp3.constants import (
 from pulp_smash.tests.pulp3.file.api_v3.utils import gen_publisher
 from pulp_smash.tests.pulp3.file.utils import populate_pulp
 from pulp_smash.tests.pulp3.file.utils import set_up_module as setUpModule  # pylint:disable=unused-import
-from pulp_smash.tests.pulp3.pulpcore.utils import gen_distribution, gen_repo
 from pulp_smash.tests.pulp3.utils import (
+    gen_distribution,
+    gen_repo,
     get_added_content,
     get_auth,
     get_versions,
