+++ pulp_smash/tests/pulp3/file/api_v3/test_crd_publications.py
@@ -16,9 +16,10 @@ from pulp_smash.tests.pulp3.constants import (
 )
 from pulp_smash.tests.pulp3.file.api_v3.utils import gen_publisher
 from pulp_smash.tests.pulp3.file.utils import set_up_module as setUpModule  # pylint:disable=unused-import
-from pulp_smash.tests.pulp3.pulpcore.utils import gen_distribution, gen_repo
 from pulp_smash.tests.pulp3.utils import (
+    gen_distribution,
     gen_remote,
+    gen_repo,
     get_auth,
     publish,
     sync,
