+++ pulp_smash/tests/pulp3/file/api_v3/test_crud_content_unit.py
@@ -15,10 +15,10 @@ from pulp_smash.tests.pulp3.constants import (
     REPO_PATH,
 )
 from pulp_smash.tests.pulp3.file.utils import set_up_module as setUpModule  # pylint:disable=unused-import
-from pulp_smash.tests.pulp3.pulpcore.utils import gen_repo
 from pulp_smash.tests.pulp3.utils import (
     delete_orphans,
     gen_remote,
+    gen_repo,
     get_auth,
     get_content,
     sync,
